+++ NEWS.md
@@ -2,6 +2,9 @@
 ## New classes and methods
 * Add `export()` to create a Zip archive of all results in CSV format.
 
+## Internals
+* Compute the position and draw only the labels in the plotting region.
+
 # dimensio 0.9.0
 ## New classes and methods
 * Add `pcoa()` to compute principal coordinates analysis.